fix: refactored gamecard to use variables and stylelint rules
